fix: move cloud extension into seaprate folder
